// frontend/src/app/core/api/model/vendorDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface VendorDto { 
    id?: number;
    vendorName: string;
    taxId: string;
    communicationScore?: number;
    deliveryScore?: number;
    qualityScore?: number;
    overallScore?: number;
    description: string;
    primaryContactName?: string;
    primaryContactEmail?: string;
    updatedAt?: string;
    createdAt?: string;
    deletedAt?: string;
}